requirements.txt
playwright>=1.47.0
pydantic>=2.7.0
pandas>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0.0
tenacity>=8.2.0
typer>=0.12.0
rich>=13.7.0
python-dateutil>=2.9.0